refactor: lift playSong up
