+++ src/components/playlist/PlaylistList.tsx
@@ -3,6 +3,7 @@ import { View } from 'react-native';
 import { FlashList } from '@shopify/flash-list';
 import Snackbar from 'react-native-snackbar';
 import { IconButton, Text } from 'react-native-paper';
+import TrackPlayer from 'react-native-track-player';
 import { Dimensions } from 'react-native';
 import { styles } from '../style';
 import SongInfo from './SongInfo';
@@ -12,6 +13,8 @@ import Song from '../../objects/SongInterface';
 import PlaylistInfo from './PlaylistInfo';
 import PlaylistMenuButton from '../buttons/PlaylistMenuButton';
 import { updateSubscribeFavList } from '../../utils/BiliSubscribe';
+import { songlistToTracklist } from '../../objects/Playlist';
+import { NoxRepeatMode } from '../player/enums/repeatMode';
 
 /*
 import Song, { dummySong } from '../../objects/SongInterface';
@@ -22,7 +25,13 @@ const DUMMYDATA = [...Array(1222).keys()].reduce(
 */
 
 export default () => {
+  const currentPlayingList = useNoxSetting(state => state.currentPlayingList);
+  const playmode = useNoxSetting(state => state.playerRepeat); // performance drain?
+  const setCurrentPlayingList = useNoxSetting(
+    state => state.setCurrentPlayingList
+  );
   const currentPlayingId = useNoxSetting(state => state.currentPlayingId);
+  const setCurrentPlayingId = useNoxSetting(state => state.setCurrentPlayingId);
   const currentPlaylist = useNoxSetting(state => state.currentPlaylist);
   const playlistShouldReRender = useNoxSetting(
     state => state.playlistShouldReRender
@@ -118,6 +127,41 @@ export default () => {
     setSearching(true);
   };
 
+  // TODO: useCallback? [currentPlaylist, currentPlayingList, playMode, index]
+  // TODO: can i somehow shove most of these into an async promise, then
+  // use a boolean flag to make a loading screen?
+  const playSong = async (song: Song) => {
+    const skipNPlay = (index: number) => {
+      TrackPlayer.skip(index).then(() => TrackPlayer.play());
+    };
+
+    const reloadPlaylistAndPlay = () => {
+      // true? currentPlaylist.songList : currentRows
+      let tracks = songlistToTracklist(currentPlaylist.songList);
+      if (playmode === NoxRepeatMode.SHUFFLE) {
+        tracks = [...tracks].sort(() => Math.random() - 0.5);
+      }
+      TrackPlayer.setQueue(tracks).then(() =>
+        skipNPlay(tracks.findIndex(track => track.song.id === song.id))
+      );
+    };
+
+    setCurrentPlayingId(song.id);
+    if (currentPlaylist.id !== currentPlayingList) {
+      setCurrentPlayingList(currentPlaylist.id);
+      reloadPlaylistAndPlay();
+    } else {
+      TrackPlayer.getQueue().then(tracks => {
+        const trackIndex = tracks.findIndex(track => track.song.id === song.id);
+        if (trackIndex === -1) {
+          reloadPlaylistAndPlay();
+        } else {
+          skipNPlay(trackIndex);
+        }
+      });
+    }
+  };
+
   /**
    * playlistShouldReRender is a global state that indicates playlist should be
    * refreshed. right now its only called when the playlist is updated in updatePlaylist.
@@ -195,6 +239,7 @@ export default () => {
               item={item}
               index={index}
               currentPlaying={item.id === currentPlayingId}
+              playSong={playSong}
               checking={checking}
               checkedProp={selected[getSongIndex(item, index)]}
               onChecked={() => toggleSelected(getSongIndex(item, index))}
