+++ src/components/playlist/SongInfo.tsx
@@ -1,18 +1,16 @@
 import * as React from 'react';
 import { Checkbox, IconButton, Text } from 'react-native-paper';
 import { View, Pressable, GestureResponderEvent } from 'react-native';
-import TrackPlayer from 'react-native-track-player';
 import { useNoxSetting } from '../../hooks/useSetting';
 import { styles } from '../style';
 import Song from '../../objects/SongInterface';
 import { seconds2MMSS } from '../../utils/Utils';
-import { playlistToTracklist } from '../../objects/Playlist';
-import { NoxRepeatMode } from '../player/enums/repeatMode';
 
 function SongInfo({
   item,
   index,
   currentPlaying,
+  playSong,
   checking = false,
   checkedProp = false,
   onChecked = () => void 0,
@@ -20,83 +18,19 @@ function SongInfo({
   item: Song;
   index: number;
   currentPlaying: boolean;
+  playSong: (song: Song) => void;
   checking?: boolean;
   checkedProp?: boolean;
   onChecked?: () => void;
 }) {
   const [title, id, artist] = [item.parsedName, item.id, item.singer];
-  const setCurrentPlayingId = useNoxSetting(state => state.setCurrentPlayingId);
   const currentPlaylist = useNoxSetting(state => state.currentPlaylist);
-  const currentPlayingList = useNoxSetting(state => state.currentPlayingList);
-  const setCurrentPlayingList = useNoxSetting(
-    state => state.setCurrentPlayingList
-  );
-  const playmode = useNoxSetting(state => state.playerRepeat); // performance drain?
   const setSongMenuCoords = useNoxSetting(state => state.setSongMenuCoords);
   const setSongMenuVisible = useNoxSetting(state => state.setSongMenuVisible);
   const setSongMenuSongIndexes = useNoxSetting(
     state => state.setSongMenuSongIndexes
   );
   const [checked, setChecked] = React.useState(false);
-
-  // TODO: useCallback? [currentPlaylist, currentPlayingList, playMode, index]
-  // TODO: can i somehow shove most of these into an async promise, then
-  // use a boolean flag to make a loading screen?
-  const playSong = async () => {
-    const reloadPlaylistAndPlay = () => {
-      let tracks = playlistToTracklist(currentPlaylist, index);
-      if (playmode === NoxRepeatMode.SHUFFLE) {
-        const currentTrack = tracks[index];
-        tracks = [...tracks].sort(() => Math.random() - 0.5);
-        TrackPlayer.setQueue(tracks).then(() => {
-          TrackPlayer.skip(tracks.indexOf(currentTrack)).then(() =>
-            TrackPlayer.play()
-          );
-        });
-      } else {
-        TrackPlayer.setQueue(tracks).then(() => {
-          TrackPlayer.skip(index).then(() => TrackPlayer.play());
-        });
-      }
-    };
-
-    setCurrentPlayingId(id);
-    if (currentPlaylist.id !== currentPlayingList) {
-      setCurrentPlayingList(currentPlaylist.id);
-      reloadPlaylistAndPlay();
-    } else {
-      TrackPlayer.getQueue().then(tracks => {
-        const trackIndex = tracks.findIndex(track => track.song.id === id);
-        if (trackIndex === -1) {
-          reloadPlaylistAndPlay();
-        } else {
-          TrackPlayer.skip(trackIndex).then(() => TrackPlayer.play());
-        }
-      });
-    }
-    /**
-       * ugly code testing out uninterrupted playlist queue change:
-      TrackPlayer.getActiveTrackIndex().then(activeTrackIndex => {
-        TrackPlayer.getQueue().then(queue => {
-          const activeTrack = queue[activeTrackIndex!];
-          let removeTrackIndex = [...Array(queue.length).keys()];
-          removeTrackIndex.splice(activeTrackIndex!, 1);
-          console.debug(removeTrackIndex, queue, activeTrack, activeTrackIndex!);
-          TrackPlayer.remove(removeTrackIndex).then(() => {
-            TrackPlayer.getQueue().then(newQueue =>
-              console.log('newQueue b4 insert', newQueue)
-            );
-            TrackPlayer.add(trackList).then(() => {
-              TrackPlayer.getQueue().then(newQueue =>
-                console.log('newQueue ', newQueue)
-              );
-            });
-          });
-        });
-      });
-      return;
-       */
-  };
 
   const toggleCheck = () => {
     setChecked(val => !val);
@@ -134,7 +68,7 @@ function SongInfo({
         </View>
       )}
       <View style={{ flex: 4.9 }}>
-        <Pressable onPress={checking ? toggleCheck : playSong}>
+        <Pressable onPress={checking ? toggleCheck : () => playSong(item)}>
           <Text variant="titleMedium">{`${String(index + 1)}. ${title}`}</Text>
           <Text variant="titleSmall" style={{ color: 'grey' }}>
             {artist}
